backend/requirements.txt
aixplain>=0.2.27
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pdfminer.six>=20221105
httpx>=0.27.0
beautifulsoup4==4.14.3